refactor search and column filters
