+++ app/actions/sites.js
@@ -4,12 +4,13 @@ import {
   GET_SITES_LOCATIONS,
   GET_SITES_SPECIES,
   GET_SITES_STATS,
-  SET_SITES_COLUMN_FILTER,
+  SET_COLUMN_FILTER,
+  SET_SEARCH_FILTER,
   SET_SITES_PARAMS,
-  SET_SITES_SEARCH,
   SET_SORT,
   SET_VIEW_MODE
 } from 'constants/action-types';
+import { TABLES } from 'constants/tables';
 import { RESULTS_PER_PAGE } from 'constants/config';
 import { push } from 'react-router-redux';
 
@@ -104,14 +105,14 @@ export function getSitesSpecies(id, type) {
 
 export function setSearchFilter(search) {
   return {
-    type: SET_SITES_SEARCH,
+    type: `${SET_SEARCH_FILTER}_${TABLES.SITES}`,
     payload: search
   };
 }
 
 export function resetSearchFilter() {
   return {
-    type: SET_SITES_SEARCH,
+    type: `${SET_SEARCH_FILTER}_${TABLES.SITES}`,
     payload: ''
   };
 }
@@ -125,14 +126,14 @@ export function setViewMode(viewMode) {
 
 export function setSitesTableSort(sort) {
   return {
-    type: `${SET_SORT}_sites`,
+    type: `${SET_SORT}_${TABLES.SITES}`,
     payload: sort
   };
 }
 
 export function setSitesTableFilter(filter) {
   return {
-    type: SET_SITES_COLUMN_FILTER,
+    type: `${SET_COLUMN_FILTER}_${TABLES.SITES}`,
     payload: filter
   };
 }
