+++ app/constants/action-types.js
@@ -1,6 +1,8 @@
-// COMMON
+// COMMON TABLE
 export const CHANGE_COLUMN_ACTIVATION = 'CHANGE_COLUMN_ACTIVATION';
 export const SET_SORT = 'SET_SORT';
+export const SET_SEARCH_FILTER = 'SET_SEARCH_FILTER';
+export const SET_COLUMN_FILTER = 'SET_COLUMN_FILTER';
 
 // COUNTRIES
 export const GET_COUNTRIES_LIST = 'GET_COUNTRIES_LIST';
@@ -13,9 +15,7 @@ export const GET_COUNTRIES_POPULATIONS = 'GET_COUNTRIES_POPULATIONS';
 export const GET_COUNTRIES_SIMILAR_SPECIES = 'GET_COUNTRIES_SIMILAR_SPECIES';
 export const GET_COUNTRIES_GEOM = 'GET_COUNTRIES_GEOM';
 export const SET_COUNTRY_PARAMS = 'SET_COUNTRY_PARAMS';
-export const SET_COUNTRY_SEARCH = 'SET_COUNTRY_SEARCH';
 export const TOGGLE_COUNTRIES_LAYER = 'TOGGLE_COUNTRIES_LAYER';
-export const SET_COUNTRY_COLUMN_FILTER = 'SET_COUNTRY_COLUMN_FILTER';
 
 // SITES
 export const CLEAR_SITES_LIST = 'CLEAR_SITES_LIST';
@@ -24,9 +24,7 @@ export const GET_SITES_STATS = 'GET_SITES_STATS';
 export const GET_SITES_LIST = 'GET_SITES_LIST';
 export const GET_SITES_SPECIES = 'GET_SITES_SPECIES';
 export const SET_SITES_PARAMS = 'SET_SITES_PARAMS';
-export const SET_SITES_SEARCH = 'SET_SITES_SEARCH';
 export const SET_VIEW_MODE = 'SET_VIEW_MODE';
-export const SET_SITES_COLUMN_FILTER = 'SET_SITES_COLUMN_FILTER';
 
 // SPECIES
 export const GET_SPECIES_STATS = 'GET_SPECIES_STATS';
@@ -38,20 +36,15 @@ export const GET_SPECIES_LOOK_ALIKE_SPECIES = 'GET_SPECIES_LOOK_ALIKE_SPECIES';
 export const GET_SPECIES_LOOK_ALIKE_SPECIES_POPULATION = 'GET_SPECIES_LOOK_ALIKE_SPECIES_POPULATION';
 export const SET_SPECIES_PARAMS = 'SET_SPECIES_PARAMS';
 export const SET_SPECIES_DETAIL_PARAMS = 'SET_SPECIES_DETAIL_PARAMS';
-export const SET_SPECIES_DETAIL_SEARCH = 'SET_SPECIES_DETAIL_SEARCH';
 export const TOGGLE_SPECIES_LAYER = 'TOGGLE_SPECIES_LAYER';
 export const TOGGLE_LEGEND_ITEM = 'TOGGLE_LEGEND_ITEM';
-export const SET_SPECIES_COLUMN_FILTER = 'SET_SPECIES_COLUMN_FILTER';
 export const SELECT_SPECIES_TABLE_ITEM = 'SELECT_SPECIES_TABLE_ITEM';
 
 // THRESHOLD
 export const SET_THRESHOLD_DATA = 'SET_THRESHOLD_DATA';
 export const SET_THRESHOLD_POSITION = 'SET_THRESHOLD_POSITION';
-export const SET_THRESHOLD_COLUMN_FILTER = 'SET_THRESHOLD_COLUMN_FILTER';
-export const SET_THRESHOLD_SEARCH_FILTER = 'SET_THRESHOLD_SEARCH_FILTER';
 
 // ADVANCED SET_SITES_SEARCH
 export const GET_SEARCH_OPTIONS = 'GET_SEARCH_OPTIONS';
 export const GET_SEARCH_RESULTS = 'GET_SEARCH_RESULTS';
-export const SET_SEARCH_FILTER = 'SET_SEARCH_FILTER';
 export const BEFORE_GET_SEARCH_RESULTS = 'BEFORE_GET_SEARCH_RESULTS';
