--- app/constants/thresholds.js
@@ -1,8 +1,0 @@
-export const ALL_THRESHOLD_COLUMNS = [
-  'scientific_name', 'english_name', 'iucn_category',
-  'population', 'a', 'b', 'c', 'caf_action_plan', 'eu_birds_directive',
-  'flyway_range', 'year_start', 'year_end', 'size_min', 'size_max',
-  'ramsar_criterion'
-];
-
-export const DEFAULT_THRESHOLD_COLUMNS = ['scientific_name', 'iucn_category', 'population', 'a', 'b', 'c', 'ramsar_criterion'];
