+++ app/containers/advanced-search/TableListHeader.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import TableListHeader from 'components/tables/TableListHeader';
 import { setSearchTableSort } from 'actions/search';
 import { changeColumnActivation } from 'actions/common';
+import { TABLES } from 'constants/tables';
 
 const mapStateToProps = (state) => ({
   selectedCategory: state.search.selectedCategory,
@@ -11,7 +12,7 @@ const mapStateToProps = (state) => ({
 const mapDispatchToProps = (dispatch) => ({
   sortBy: (sort) => dispatch(setSearchTableSort(sort)),
   filterBy: () => {},
-  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, 'search'))
+  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, TABLES.SEARCH))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(TableListHeader);
