+++ app/containers/countries/TableListHeader.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import TableListHeader from 'components/tables/TableListHeader';
 import { setCountriesTableSort, setCountriesTableFilter } from 'actions/countries';
 import { changeColumnActivation } from 'actions/common';
+import { TABLES } from 'constants/tables';
 
 const mapStateToProps = (state, ownProps) => ({
   selectedCategory: ownProps.selectedCategory || state.countries.selectedCategory,
@@ -12,7 +13,7 @@ const mapStateToProps = (state, ownProps) => ({
 const mapDispatchToProps = (dispatch) => ({
   sortBy: (sort) => dispatch(setCountriesTableSort(sort)),
   filterBy: (filter) => dispatch(setCountriesTableFilter(filter)),
-  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, 'countries'))
+  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, TABLES.COUNTRIES))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(TableListHeader);
