+++ app/containers/species/TableListHeader.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import TableListHeader from 'components/tables/TableListHeader';
 import { setSpeciesTableSort, setSpeciesFilter } from 'actions/species';
 import { changeColumnActivation } from 'actions/common';
+import { TABLES } from 'constants/tables';
 
 const mapStateToProps = (state) => ({
   selectedCategory: state.species.selectedCategory,
@@ -11,7 +12,7 @@ const mapStateToProps = (state) => ({
 const mapDispatchToProps = (dispatch) => ({
   sortBy: (sort) => dispatch(setSpeciesTableSort(sort)),
   filterBy: (filter) => dispatch(setSpeciesFilter(filter)),
-  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, 'species'))
+  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, TABLES.SPECIES))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(TableListHeader);
