+++ app/containers/threshold/TableListHeader.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux';
 import TableListHeader from 'components/tables/TableListHeader';
 import { setThresholdTableSort, setThresholdColumnFilter } from 'actions/threshold';
 import { changeColumnActivation } from 'actions/common';
+import { TABLES } from 'constants/tables';
 
 const mapStateToProps = (state) => ({
   sort: state.threshold.sort
@@ -10,7 +11,7 @@ const mapStateToProps = (state) => ({
 const mapDispatchToProps = (dispatch) => ({
   sortBy: (sort) => dispatch(setThresholdTableSort(sort)),
   filterBy: (filter) => dispatch(setThresholdColumnFilter(filter)),
-  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, 'threshold'))
+  changeColumnActivation: (column) => dispatch(changeColumnActivation(column, TABLES.THRESHOLDS))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(TableListHeader);
